Web Bundler Common is now an extension
