+++ common/deployment/pom.xml
@@ -4,12 +4,17 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>io.quarkiverse.web-bundler</groupId>
-        <artifactId>quarkus-web-bundler-parent</artifactId>
+    <artifactId>quarkus-web-bundler-common-parent</artifactId>
         <version>999-SNAPSHOT</version>
     </parent>
     <artifactId>quarkus-web-bundler-common-deployment</artifactId>
     <name>Quarkus Web Bundler - Common - Deployment</name>
     <dependencies>
+    <dependency>
+      <groupId>io.quarkiverse.web-bundler</groupId>
+      <artifactId>quarkus-web-bundler-common</artifactId>
+      <version>${project.version}</version>
+    </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-arc-deployment</artifactId>
