+++ common/deployment/src/main/java/io/quarkiverse/web/bundler/deployment/ConfiguredEntryPoint.java
