+++ common/deployment/src/main/java/io/quarkiverse/web/bundler/deployment/WebBundlerConfig.java
