+++ common/deployment/src/main/java/io/quarkiverse/web/bundler/deployment/items/ExternalFileWebAsset.java
