+++ common/deployment/src/main/java/io/quarkiverse/web/bundler/deployment/items/GeneratedWebResourceBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.web.bundler.deployment.web;
+package io.quarkiverse.web.bundler.deployment.items;
 
 import java.nio.file.Path;
