+++ common/deployment/src/main/java/io/quarkiverse/web/bundler/deployment/items/StaticAssetsBuildItem.java
@@ -1,0 +1,10 @@
+package io.quarkiverse.web.bundler.deployment.items;
+
+import java.util.List;
+
+public final class StaticAssetsBuildItem extends WebAssetsBuildItem {
+
+    public StaticAssetsBuildItem(List<WebAsset> webAssets) {
+        super(webAssets);
+    }
+}
