+++ common/deployment/src/main/java/io/quarkiverse/web/bundler/deployment/util/PathUtils.java
