+++ common/pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.quarkiverse.web-bundler</groupId>
+        <artifactId>quarkus-web-bundler-parent</artifactId>
+        <version>999-SNAPSHOT</version>
+    </parent>
+    <artifactId>quarkus-web-bundler-common-parent</artifactId>
+    <name>Quarkus Web Bundler - Common - Parent</name>
+    <packaging>pom</packaging>
+    <modules>
+        <module>runtime</module>
+        <module>deployment</module>
+    </modules>
+</project>
