+++ deployment/src/main/java/io/quarkiverse/web/bundler/deployment/BundlingProcessor.java
@@ -20,9 +20,9 @@ import io.mvnpm.esbuild.Bundler;
 import io.mvnpm.esbuild.model.BundleResult;
 import io.quarkiverse.web.bundler.deployment.items.GeneratedBundleBuildItem;
 import io.quarkiverse.web.bundler.deployment.items.GeneratedEntryPointBuildItem;
+import io.quarkiverse.web.bundler.deployment.items.GeneratedWebResourceBuildItem;
+import io.quarkiverse.web.bundler.deployment.items.GeneratedWebResourceBuildItem.SourceType;
 import io.quarkiverse.web.bundler.deployment.items.ReadyForBundlingBuildItem;
-import io.quarkiverse.web.bundler.deployment.web.GeneratedWebResourceBuildItem;
-import io.quarkiverse.web.bundler.deployment.web.GeneratedWebResourceBuildItem.SourceType;
 import io.quarkiverse.web.bundler.runtime.Bundle;
 import io.quarkiverse.web.bundler.runtime.BundleRedirectHandlerRecorder;
 import io.quarkiverse.web.bundler.runtime.WebBundlerBuildRecorder;
