+++ deployment/src/main/java/io/quarkiverse/web/bundler/deployment/devui/WebBundlerDevUIProcessor.java
@@ -12,8 +12,8 @@ import java.util.stream.Collectors;
 import io.quarkiverse.web.bundler.deployment.WebBundlerConfig;
 import io.quarkiverse.web.bundler.deployment.items.EntryPointBuildItem;
 import io.quarkiverse.web.bundler.deployment.items.GeneratedEntryPointBuildItem;
+import io.quarkiverse.web.bundler.deployment.items.GeneratedWebResourceBuildItem;
 import io.quarkiverse.web.bundler.deployment.items.WebDependenciesBuildItem;
-import io.quarkiverse.web.bundler.deployment.web.GeneratedWebResourceBuildItem;
 import io.quarkus.deployment.IsDevelopment;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
