+++ docs/modules/ROOT/pages/advanced-guides.adoc
@@ -416,3 +416,4 @@ or in a Qute template:
 {inject:bundle.script("app")}
 {inject:bundle.style("app")}
 ----
+
