+++ runtime/src/main/java/io/quarkiverse/web/bundler/runtime/WebBundlerBuildRecorder.java
@@ -28,5 +28,4 @@ public class WebBundlerBuildRecorder {
 
         };
     }
-
 }
